+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changed
 
+- [#250][]: Replaced symbolic errors with Rails 5-style detailed errors.
 - [#269][]: Prevented proc defaults from being eagerly evaluated.
 - [#264][]: Renamed `model` filter to `object`.
 - [#213][]: Remove transaction support. Database transactions will need to be
@@ -507,5 +508,6 @@
   [#239]: https://github.com/orgsync/active_interaction/issues/239
   [#244]: https://github.com/orgsync/active_interaction/issues/244
   [#248]: https://github.com/orgsync/active_interaction/issues/248
+  [#250]: https://github.com/orgsync/active_interaction/issues/250
   [#264]: https://github.com/orgsync/active_interaction/issues/264
   [#269]: https://github.com/orgsync/active_interaction/issues/269